[Justice Counts] Adding new READ_ONLY role (Recidiviz/recidiviz-data#21068)
## Description of the change

We'd like to introduce a new READ_ONLY role to the AgencyTeamMemberRole
enum. Users with this role should be able to view everything in their
agency, but should not be able to make any changes. That means all text
boxes, radio buttons, etc. should be disabled for them. We will be
granting this role to CSG folks on Production so they can view agencies
but not accidentally make changes.

Ivan is working on the FE portion of this task.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Contributes to
https://app.zenhub.com/workspaces/justice-counts--622109ea3bb6e0001d2d06ad/issues/gh/recidiviz/justice-counts/647

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 23e0b79624edd57fb1220eba25405184d3208b92