[Justice Counts] populate is_report_datapoint and source_id in code (Recidiviz/recidiviz-data#20965)
## Description of the change

This PR makes 2 'forward thinking' changes:
1. Sets Datapoint.is_report_datapoint to indicate if a new datapoint
belongs to a report (True), or is an agency datapoint (False)
2. Populates Datapoint.source_id for all new datapoints, so that all new
report datapoints start getting this set appropriately (old report
datapoints will still have this as null)

Once this PR is merged, the db will be in a weird state. New datapoints
will have the "correct" values for is_report_datapoint and source_id,
but old ones will not.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#20909

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
GitOrigin-RevId: 9c6af06d4a4af0b24c6e1aaa3d023bce56e49b7b