[Justice Counts] Add is_superagency column and send to FE (Recidiviz/recidiviz-data#21087)
## Description of the change

This is a bit of a strange PR and it does a few things. The main
motivation was that in order to support some upcoming changes (e.g.
[this
task](https://app.zenhub.com/workspaces/justice-counts--622109ea3bb6e0001d2d06ad/issues/gh/recidiviz/justice-counts/607))
the FE is going to need to know whether an agency is a superagency or
not. Currently, we determine whether an agency is a superagency by
querying the DB to find out if it has any child agencies. This is fine,
but I wanted to avoid having to make this extra DB call all the time,
since most agencies _aren't_ superagencies, so it's kind of a waste for
them. So this led me to think that we can just add a boolean
`is_superagency` column that we set to true/false, and then we know as
soon as we load an agency from the DB whether it's a superagency or not,
and we only need to go back to the DB to fetch child agencies if it
truly is a superagency.

After making that change, I realized I needed to update the admin panel
so we can set this flag. I changed the Superagency Provisioning page so
that it lets you set this flag, and also filtered the table there to
only show agencies with the flag set to True. See video:

https://github.com/Recidiviz/recidiviz-data/assets/99899591/42600811-c5b7-4356-9e41-aded4065cde6

I'm also passing `is_superagency` and `child_agencies` to the FE now,
and I verified that the FE can access them:
<img width="678" alt="Screenshot 2023-05-24 at 6 34 04 PM"
src="https://github.com/Recidiviz/recidiviz-data/assets/99899591/43928425-a740-4525-8018-a77c83b5da5e">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Contributes to
https://app.zenhub.com/workspaces/justice-counts--622109ea3bb6e0001d2d06ad/issues/gh/recidiviz/justice-counts/607

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 27321a9f5c3f844e8d0d2292c0f0a4275760d0de