[Ingest View Tests] Migrate tests that were failing due to datetime formatting (Recidiviz/recidiviz-data#21131)
## Description of the change

This updates the emulator to the latest version, and migrates tests that
were broken due to datetime formatting.

Note: It looks like postgres was more lenient with datetime parsing, so
some strings would get parsed as datetimes even if they didn't actually
match any of the formats that we try when parsing raw data fields with
`is_datetime` set. BigQuery is not lenient here, so these dates stay as
strings the whole way through. Some of the expected results had to be
updated to match this.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

most of Recidiviz/recidiviz-data#20960

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7a957988a9f2c68144e2371f67a5cfaf4db929e6