Updating state_code_branch to have a DummyOperator in between (Recidiviz/recidiviz-data#21148)
## Description of the change

We saw that a RuntimeError occurred when we tried to pass task group ids
to the branch operator when it was actually expecting task ids. This was
not able to be caught by a unit test, because it all compiles fine 🙃.
This adds an EmptyOperator before the task branch in order to soothe
runtime errors. See
[link](https://stackoverflow.com/questions/72306836/airflow-branch-operator-and-task-group-invalid-task-ids).

## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c22c259b3e6287661d164ad4e5131a4efc4b03d7