[US_MI] Gate incarceration periods revisions to staging for now (Recidiviz/recidiviz-data#21211)
## Description of the change

At the end of last week, I merged in a
[PR](https://github.com/Recidiviz/recidiviz-data/pull/20702) for
incarceration periods revisions and started a partial rerun for them in
staging. The rerun crashed over the weekend due to a enum mapping error
(which I fixed
[here](https://github.com/Recidiviz/recidiviz-data/pull/21163)) which
means it won't finish before the next prod deploy as planned. Since we
won't have a chance to fully test the revisions before it gets into
prod, we want to gate the revisions to staging only for now.

This PR:
- combines all the revisions from the two PRs listed above into a _v3
version of incarceration periods
- restores the previous _v2 version of incarceration periods
- gates the _v3 version of incarceration periods to staging only

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b785608ba62c3df73aae15d10ef9d4a2390d86bb