[Ingest] Use milliseconds in bounds when generating ingest view results (Recidiviz/recidiviz-data#21232)
## Description of the change

We generate ingest view results incrementally, by picking boundary
timestamps and then running ingest view queries that include all data up
to that boundary, diffing it with results of the prior boundary, and
just writing the difference.

A recent change (https://github.com/Recidiviz/recidiviz-data/pull/19257)
updated how we add these timestamps into the query, and in the process
started dropping microseconds from the timestamps. This means that if
the most recent raw data had a timestamp of "2023-05-31
10:07:05.839228", we would use a bound of "2023-05-31 10:07:05" and
accidentally exclude that file, so only prior data is included for that
file.

This is exacerbated with raw data pruning. If ingest view results are
generated right after raw data pruning was run, the next most recent
file will have been deleted, so instead of only including raw data
through that file, it will only include the initial version of a raw
data file which could be 1+ years old.

This PR updates so that we include microseconds in the datetime bound
clause for the upper bound query, but do NOT include microseconds in the
following places:
* Temporarily, in the lower bound query so that we can ensure we don't
generate a gap in ingest view results after pushing this change.
* In any clause that queries metadata date columns in the ingest view
results tables (these have always had microseconds truncated).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Fix for Recidiviz/recidiviz-data#21230

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ba4ed6a1b4f64c10895803aa3c4e7c5d99e29b1d