[US_PA] Fix state staff caseload type period view (Recidiviz/recidiviz-data#21277)
## Description of the change

In the first version of `supervision_staff_caseload_type_period` view, I
incorrectly assumed that each staff member appears exactly one time in
the static data we have from PA about specialized agents. Once the view
was deployed I realized this was not the case, and the view logic needed
a little more nuance.

It appears that we have received the static table of specialized agents
from PA exactly two times - once on 2022-08-16, and once on 2023-03-02.
Since these dates only indicate when we received the data from the
state, and not when an agent's caseload type actually changed, I am not
using them to construct specialized caseload type periods. Instead, I am
assuming the most recent update to an agent's specialized caseload is
the most accurate; if the caseload type in the previous transfer was
different, the previous caseload type is disregarded, since we do not
know when the agent switched from one to another.

In the case where more than one `AgentType` was associated with the same
`external_id` in the latest transfer, we store the one that is different
from the one in the previous transfer, if possible. This is based on the
assumption that the different value would not be included if the agent
hadn't had a change to their `AgentType`. If `AgentType` values are
identical over time, we randomly select one that will be identical every
time ingest view results are generated.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8d06610a8a2d89ff971957fae36fa26434a29013